Rename test class
